refactor(api): move eln import into services/eln
The ELN import will grow a translator per source publisher (SciLog,
openBIS, ...). Group the ELN modules under services/eln, with the
publisher translators in a translators/ submodule, so the family has a
single home and the top-level services/ directory stays legible.

Pure relocation: files moved via git mv and the SciLog translator class
renamed ElnTranslator -> ScilogTranslator. No behaviour change.

See: #669
