+++ api/src/services/index.ts
@@ -6,5 +6,5 @@
 export * from './ro-crate-export.service';
 export * from './entity-builder.service';
 export * from './archive.service';
-export * from './eln-import.service';
+export * from './eln/import.service';
 export * from './file-storage.service';
